# api/openapi.yml
openapi: 3.1.0
info:
  title: SwiftUI View Factory API
  description: |
    A standalone service that converts UI mockups or structured layout trees
    into production-grade SwiftUI views, optionally wired to backend SDKs.

    The API supports a curated subset of SwiftUI components. Supported
    ``LayoutNode.type`` values are:

    - ``VStack``
    - ``HStack``
    - ``Text``
    - ``Image``
    - ``Button``
    - ``Spacer``
    - ``ScrollView``
    - ``ZStack``
    - ``Conditional``
    - ``TextField``
    - ``Form``
    - ``List``
    - ``Section``
    - ``NavigationStack``
  version: 1.2.0

servers:
  # The FastAPI server is mounted at the root path when run locally, so the specification should reflect the unversioned base URL.
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.fountain.coach/v1
    description: Production v1

tags:
  - name: Factory
    description: Convert mockups and layout trees into SwiftUI views
  - name: Secrets
    description: Internal endpoints for managing API secrets

security:
  - BearerAuth: []

paths:
  /factory/interpret:
    post:
      summary: Interpret UI mockup image
      description: Upload a UI mock or sketch to generate a structured layout tree.
      tags:
        - Factory
      operationId: interpretLayout
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Structured layout response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutInterpretationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /factory/generate:
    post:
      summary: Generate SwiftUI view code
      description: Converts a structured layout into a SwiftUI `View` struct with optional backend wiring.
      tags:
        - Factory
      operationId: generateSwiftUIView
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                layout:
                  $ref: '#/components/schemas/LayoutNode'
                name:
                  type: string
                  description: Optional name for the generated SwiftUI view (e.g., HomeView)
                style:
                  type: object
                  properties:
                    indent:
                      type: integer
                      default: 2
                    header_comment:
                      type: boolean
                      default: true
                    font:
                      type: string
                      description: Font applied to Text and Button views
                    color:
                      type: string
                      description: Foreground color name
                    spacing:
                      type: integer
                      description: Spacing value for stacks
                    bold:
                      type: boolean
                      description: Apply bold styling to Text and Button views
                    italic:
                      type: boolean
                      description: Apply italic styling to Text and Button views
                    padding:
                      type: integer
                      description: Padding value applied to leaf views
                    background_color:
                      type: string
                      description: Background color name
                    corner_radius:
                      type: integer
                      description: Corner radius for leaf views
                backend_hooks:
                  type: boolean
                  default: false
                  description: |
                    If true, the generated SwiftUI view includes an `.onAppear`
                    block where analytics or network logic can be invoked.
      responses:
        '200':
          description: Generated SwiftUI view code
          content:
            application/json:
              schema:
                type: object
                properties:
                  swift:
                    type: string
                    example: |
                      struct GeneratedView: View {
                          var body: some View {
                              Text("Hello")
                          }
                      }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /secret:
    get:
      summary: Retrieve the OpenAI API key
      description: |
        Returns the API key used when communicating with OpenAI. In local
        development the key is loaded from a `.env` file. In production the
        container expects the `OPENAI_API_KEY` environment variable to be set.
      tags:
        - Secrets
      operationId: getOpenAIKey
      responses:
        '200':
          description: API key response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAIKeyResponse'


components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LayoutInterpretationResponse:
      type: object
      required:
        - structured
      properties:
        structured:
          $ref: '#/components/schemas/LayoutNode'
        description:
          type: string
          description: Optional natural language summary
        version:
          type: string
          example: "layout-v1"
        log:
          type: string
          description: Raw communication log between the service and OpenAI
      example:
        structured:
          type: VStack
          children:
            - type: Text
              text: Hello
        description: Simple VStack with Hello text
        version: layout-v1

    LayoutNode:
      type: object
      required:
        - type
      properties:
        id:
          type: string
          description: A unique node identifier
        role:
          type: string
          description: Semantic role (e.g., "header", "submit")
          nullable: true
        tag:
          type: string
          description: Developer hint or custom logic
          nullable: true
        type:
          type: string
          description: SwiftUI component type
          enum:
            - VStack
            - HStack
            - Text
            - Image
            - Button
            - Spacer
            - ScrollView
            - ZStack
            - Conditional
            - TextField
            - Form
            - List
            - Section
            - NavigationStack
        text:
          type: string
          nullable: true
        children:
          type: array
          items:
            $ref: '#/components/schemas/LayoutNode'
          nullable: true
          description: >
            Child layout nodes rendered inside container views. Multiple entries
            are typically used for ``VStack``, ``HStack``, ``ZStack``, ``Form`` or
            ``NavigationStack`` groups.
        condition:
          type: string
          description: Condition expression controlling the branch
          nullable: true
        then:
          $ref: '#/components/schemas/LayoutNode'
          nullable: true
        else:
          $ref: '#/components/schemas/LayoutNode'
          nullable: true
      example:
        type: VStack
        children:
          - type: Text
            text: Hello

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: One or more fields are invalid.
        detail:
          type: string
          nullable: true
        log:
          type: string
          nullable: true
          description: Raw communication log if available

    OpenAIKeyResponse:
      type: object
      required:
        - api_key
      properties:
        api_key:
          type: string
          example: sk-abc123